requirements.txt
# Cython==0.29.24
autopep8==2.0.1
bcrypt==4.0.1
blinker==1.5
click==8.1.3
colorama==0.4.6
coverage==5.5
cssbeautifier==1.14.7
deprecation==2.1.0
djlint==1.19.15
dnspython==2.3.0
EditorConfig==0.12.3
email-validator==1.3.1
eyed3==0.9.7
filetype==1.2.0
Flask==2.2.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
html-tag-names==0.1.2
html-void-elements==0.1.0
idna==3.4
itsdangerous==2.1.0
Jinja2==3.1.2
jsbeautifier==1.14.7
MarkupSafe==2.1.2
packaging==23.0
pathspec==0.11.0
pycodestyle==2.10.0
PyJWT==2.6.0
# PyYAML==5.4.1
regex==2022.10.31
six==1.16.0
SQLAlchemy==2.0.4
toml==0.10.2
tomli==2.0.1
tqdm==4.64.1
typing_extensions==4.5.0
Werkzeug==2.2.3
WTForms==3.0.1
pwinput==1.0.3
pillow==10.1.0